Re-factored news fetch
