+++ Open Judge System/Web/OJS.Web/Areas/Administration/Providers/Contracts/INewsProvider.cs
@@ -1,0 +1,11 @@
+﻿namespace OJS.Web.Areas.Administration.Providers.Contracts
+{
+    using System.Collections.Generic;
+
+    using OJS.Data.Models;
+
+    public interface INewsProvider
+    {
+        IEnumerable<News> FetchNews();
+    }
+}
