+++ Open Judge System/Web/OJS.Web/OJS.Web.csproj
@@ -764,6 +764,10 @@
     <Compile Include="Areas\Administration\Controllers\SubmissionsController.cs" />
     <Compile Include="Areas\Administration\Controllers\TestsController.cs" />
     <Compile Include="Areas\Administration\Controllers\UsersController.cs" />
+    <Compile Include="Areas\Administration\Providers\Common\BaseNewsProvider.cs" />
+    <Compile Include="Areas\Administration\Providers\Contracts\INewsProvider.cs" />
+    <Compile Include="Areas\Administration\Providers\InfoManNewsProvider.cs" />
+    <Compile Include="Areas\Administration\Providers\InfosNewsProvider.cs" />
     <Compile Include="Areas\Administration\ViewModels\AntiCheat\SubmissionSimilarityViewModel.cs" />
     <Compile Include="Areas\Administration\ViewModels\AntiCheat\AntiCheatByIpAdministrationViewModel.cs" />
     <Compile Include="Areas\Administration\ViewModels\AntiCheat\IpSubmissionsAdministrationViewModel.cs" />
